+++ app/uk/gov/hmrc/bforms/views/aggregates_levy.scala.html
@@ -1,0 +1,5 @@
+@()(implicit request: Request[_], messages: Messages))
+
+@uk.gov.hmrc.bforms.views.html.main_template(title = "bforms-frontend aggregates levy", bodyClasses = None) {
+    <h1>Hello from bforms-frontend aggregates levy !</h1>
+}
