--- app/uk/gov/hmrc/bforms/views/helloworld/hello_world.scala.html
@@ -1,5 +1,0 @@
-@()(implicit request: Request[_], messages: Messages))
-
-@uk.gov.hmrc.bforms.views.html.main_template(title = "Hello from bforms-frontend", bodyClasses = None) {
-    <h1>Hello from bforms-frontend !</h1>
-}
