+++ conf/app.routes
@@ -1,3 +1,3 @@
 # microservice specific routes
 
-GET        /hello-world             uk.gov.hmrc.bforms.controllers.HelloWorld.helloWorld
+GET        /aggregates-levy    uk.gov.hmrc.bforms.controllers.AggregatesLevy.aggregatesLevy
