+++ test/uk/gov/hmrc/bforms/controllers/AggregatesLevyControllerSpec.scala
@@ -1,0 +1,44 @@
+/*
+ * Copyright 2016 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package uk.gov.hmrc.bforms.controllers
+
+import play.api.http.Status
+import play.api.test.FakeRequest
+import play.api.test.Helpers._
+import uk.gov.hmrc.play.test.{UnitSpec, WithFakeApplication}
+
+
+class AggregatesLevyControllerSpec extends UnitSpec with WithFakeApplication{
+
+  val fakeRequest = FakeRequest("GET", "/")
+
+
+  "GET /aggregates-levy" should {
+    "return 200" in {
+      val result = AggregatesLevy.aggregatesLevy(fakeRequest)
+      status(result) shouldBe Status.OK
+    }
+
+    "return HTML" in {
+      val result = AggregatesLevy.aggregatesLevy(fakeRequest)
+      contentType(result) shouldBe Some("text/html")
+      charset(result) shouldBe Some("utf-8")
+      contentAsString(result) should include("aggregates levy")    }
+  }
+
+
+}
